# Trace simplification