--- a/README.markdown
+++ b/README.markdown
@@ -46,7 +46,7 @@ There are a few different packages under the Déjà Fu umbrella:
 |   | Version | Summary |
 | - | ------- | ------- |
 | [concurrency][h:conc]    | 1.4.0.2 | Typeclasses, functions, and data types for concurrency and STM. |
-| [dejafu][h:dejafu]       | 1.3.1.0 | Systematic testing for Haskell concurrency. |
+| [dejafu][h:dejafu]       | 1.3.2.0 | Systematic testing for Haskell concurrency. |
 | [hunit-dejafu][h:hunit]  | 1.1.0.2 | Deja Fu support for the HUnit test framework. |
 | [tasty-dejafu][h:tasty]  | 1.1.0.1 | Deja Fu support for the Tasty test framework. |