--- a/dejafu/LICENSE
+++ b/dejafu/LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2015--2017, Michael Walker <mike@barrucadu.co.uk>
+Copyright (c) 2015--2018, Michael Walker <mike@barrucadu.co.uk>
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the