--- a/dejafu/Test/DejaFu/Internal.hs
+++ b/dejafu/Test/DejaFu/Internal.hs
@@ -40,6 +40,7 @@ data Settings n a = Settings
   , _debugShow :: Maybe (a -> String)
   , _debugPrint :: Maybe (String -> n ())
   , _earlyExit :: Maybe (Either Failure a -> Bool)
+  , _equality :: Maybe (a -> a -> Bool)
   }
 
 -- | How to explore the possible executions of a concurrent program.