--- a/dejafu/Test/DejaFu/Schedule.hs
+++ b/dejafu/Test/DejaFu/Schedule.hs
@@ -1,6 +1,6 @@
 -- |
 -- Module      : Test.DejaFu.Schedule
--- Copyright   : (c) 2016--2017 Michael Walker
+-- Copyright   : (c) 2016--2018 Michael Walker
 -- License     : MIT
 -- Maintainer  : Michael Walker <mike@barrucadu.co.uk>
 -- Stability   : experimental