--- a/dejafu/dejafu.cabal
+++ b/dejafu/dejafu.cabal
@@ -2,7 +2,7 @@
 -- documentation, see http://haskell.org/cabal/users-guide/
 
 name:                dejafu
-version:             1.3.1.0
+version:             1.3.2.0
 synopsis:            A library for unit-testing concurrent programs.
 
 description:
@@ -20,7 +20,7 @@ license:             MIT
 license-file:        LICENSE
 author:              Michael Walker
 maintainer:          mike@barrucadu.co.uk
-copyright:           (c) 2015--2017 Michael Walker
+copyright:           (c) 2015--2018 Michael Walker
 category:            Concurrency
 build-type:          Simple
 extra-source-files:  README.markdown CHANGELOG.rst
@@ -33,7 +33,7 @@ source-repository head
 source-repository this
   type:     git
   location: https://github.com/barrucadu/dejafu.git
-  tag:      dejafu-1.3.1.0
+  tag:      dejafu-1.3.2.0
 
 library
   exposed-modules:     Test.DejaFu