… dejafu/Test/DejaFu/SCT/Internal.hs
… dejafu/Test/DejaFu/SCT/Internal/DPOR.hs
… dejafu/Test/DejaFu/Utils.hs
… doc/getting_started.rst