## Summary

Adds a "trace simplification" pass, which tries to reduce context switching in execution traces.

The core idea is that the dependency function lets us implement a (slightly conservative) semantics-preserving simplifier, *without needing to re-run the test case after every simplification pass*.  This is fundamentally different to QuickCheck-style *shrinking*, which produces a set of candidate shrinks and then evaluates the test case on them all, repeating this candidate/test process until a minimum is found.  In contrast we just rewrite a single trace into a simpler form in one go, then get dejafu to generate an execution trace from it.  So we do re-run the test case, but only once, rather than many times.

Relevant memo here: https://memo.barrucadu.co.uk/simplifying-execution-traces.html

**To do:**

- [x] Performance measurements
- [x] Prepare a release

**Related issues:** closes #183 

## Checklist

**If this is fixing a bug or adding a feature:**

- [x] Add new tests to dejafu-tests
- [x] Simplification turned on for all tests, with recoverable errors turned into fatal ones

## Benchmark results

As expected, turning on simplification without equality rather increases the memory usage.

Turning on equality appears to reduce memory usage compared to both new things disabled, but I ran the case again and got a different residency.  So I don't think there's any significant difference in memory usage in the normal case.

**Before:**

```
$ stack exec -- dejafu-tests +RTS -s

 202,035,183,224 bytes allocated in the heap                                                                                          
  54,068,387,672 bytes copied during GC
      23,383,168 bytes maximum residency (16413 sample(s))
         189,312 bytes maximum slop
              48 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     178152 colls,     0 par   104.704s  104.697s     0.0006s    0.0700s
  Gen  1     16413 colls,     0 par   10.102s  10.102s     0.0006s    0.0020s

  TASKS: 237597 (237592 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  213.823s  (206.111s elapsed)
  GC      time  114.807s  (114.799s elapsed)
  EXIT    time    0.002s  (  0.009s elapsed)
  Total   time  328.632s  (320.920s elapsed)

  Alloc rate    944,871,571 bytes per MUT second

  Productivity  65.1% of total user, 64.2% of total elapsed
```

**After (simplification + equality disabled):**

```
$ stack exec -- dejafu-tests +RTS -s

 202,910,415,960 bytes allocated in the heap                                                                                          
  54,427,078,584 bytes copied during GC
      25,937,656 bytes maximum residency (16263 sample(s))
         215,304 bytes maximum slop
              54 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     179163 colls,     0 par   105.860s  105.848s     0.0006s    0.0611s
  Gen  1     16263 colls,     0 par    9.927s   9.929s     0.0006s    0.0019s

  TASKS: 238437 (238432 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  215.786s  (208.028s elapsed)
  GC      time  115.787s  (115.777s elapsed)
  EXIT    time    0.001s  (  0.005s elapsed)
  Total   time  331.575s  (323.810s elapsed)

  Alloc rate    940,330,203 bytes per MUT second

  Productivity  65.1% of total user, 64.2% of total elapsed
```

**After (simplification enabled, equality disabled):**

```
$ stack exec -- dejafu-tests +RTS -s

 204,336,663,872 bytes allocated in the heap                                                                                          
  55,850,952,152 bytes copied during GC
      38,247,224 bytes maximum residency (16367 sample(s))
         410,824 bytes maximum slop
             106 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     180433 colls,     0 par   106.596s  106.555s     0.0006s    0.0890s
  Gen  1     16367 colls,     0 par    9.948s   9.946s     0.0006s    0.0022s

  TASKS: 237824 (237819 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  214.022s  (206.069s elapsed)
  GC      time  116.544s  (116.501s elapsed)
  EXIT    time    0.003s  (  0.011s elapsed)
  Total   time  330.569s  (322.580s elapsed)

  Alloc rate    954,747,612 bytes per MUT second

  Productivity  64.7% of total user, 63.9% of total elapsed
```

**After (simplification disabled, equality enabled):**

```
$ stack exec -- dejafu-tests +RTS -s

 203,156,926,888 bytes allocated in the heap                                                                                          
  53,891,247,904 bytes copied during GC
      14,394,344 bytes maximum residency (15018 sample(s))
         153,920 bytes maximum slop
              39 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     180700 colls,     0 par   107.043s  107.013s     0.0006s    0.0364s
  Gen  1     15018 colls,     0 par    9.369s   9.369s     0.0006s    0.0022s

  TASKS: 237315 (237310 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  216.034s  (208.465s elapsed)
  GC      time  116.412s  (116.382s elapsed)
  EXIT    time    0.001s  (  0.002s elapsed)
  Total   time  332.448s  (324.850s elapsed)

  Alloc rate    940,395,264 bytes per MUT second

  Productivity  65.0% of total user, 64.2% of total elapsed
```


**After (simplification + equality enabled):**

```
$ stack exec -- dejafu-tests +RTS -s

 203,114,989,304 bytes allocated in the heap                                                                                          
  53,830,904,480 bytes copied during GC
      14,187,552 bytes maximum residency (14961 sample(s))
         151,824 bytes maximum slop
              39 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     180717 colls,     0 par   105.581s  105.557s     0.0006s    0.0342s
  Gen  1     14961 colls,     0 par    9.210s   9.211s     0.0006s    0.0021s

  TASKS: 237358 (237353 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  212.736s  (204.671s elapsed)
  GC      time  114.791s  (114.768s elapsed)
  EXIT    time    0.002s  (  0.011s elapsed)
  Total   time  327.529s  (319.450s elapsed)

  Alloc rate    954,775,195 bytes per MUT second

  Productivity  65.0% of total user, 64.1% of total elapsed
```